Revert "merge OpenRoad?master with naveen"

+++ src/cts/src/Clustering.cpp
@@ -308,10 +308,10 @@ void Clustering::minCostFlow(const std::vector<std::pair<float, float>>& means,
     src_sink_edges.push_back(graph.addArc(src, sink));
   }
   // edges between sinks and clusters
-  std::vector<double> costs;
+  std::vector<float> costs;
   for (size_t i = 0; i < sinks_.size(); ++i) {
     for (size_t j = 0; j < means.size(); ++j) {
-      double d = calcDist(means[j], &sinks_[i]);
+      float d = calcDist(means[j], &sinks_[i]);
       if (d <= dist) {
         d = std::pow(d, power);
         if (d < std::numeric_limits<int>::max()) {
